backend/pyproject.toml
[project]
name = "shadowcut"
dynamic = ["version"]
description = "There's no description yet"
readme = "README.md"
requires-python = ">=3.12,<4.0"

[project.urls]
"Homepage" = "https://github.com/shadowyohan/shadowcut"
"Bug Reports" = "https://github.com/shadowyohan/shadowcut/issues"
"Source" = "https://github.com/shadowyohan/shadowcut"

[tool.poetry]
version = "2.0.1"

[tool.poetry.dependencies]
fastapi = "^0.115.8"
dishka = "^1.4.2"
uvicorn = "^0.34.0"
alembic = "^1.14.1"
pydantic-settings = "^2.7.1"
sqlalchemy-utils = "^0.41.2"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.37" }
asyncpg = "^0.30.0"
pytest = "^8.3.4"
poetry = "^2.0.1"
faker = "^35.2.0"
pytest-asyncio = "^0.25.3"
rust-just = "^1.39.0"

[tool.ruff]
line-length = 120
target-version = "py310"
extend-exclude = ["adapters/alembic/versions"]

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true

[tool.ruff.lint]
select = ["UP037"]

[tool.ruff.lint.flake8-copyright]
author = "giorka"

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "first-party",
    "local-folder",
    "third-party",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F403"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true
